feat: Add search history & top searches
A cherry-pick from the Gatsby's merge commit of this same feature.

* Skip rendering the history if there isn't any

* Clear the history when its button is clicked

* Stop trying to open the link in a new tab

In a previous design spec we opened the links in a new tab but we are not doing this anymore

* Remove onClear prop, does not seem to be needed

* Show the search history component

* Show up to 4 search history items max

The design spec mentions 4 max:
https://www.figma.com/file/kJywgfIx8LBIzggCY5AeDy/Handoff-v3.1-(Search)?node-id=2%3A58234

* Add the top searches

* Extract a `formatSearchPath` util function

* Use a link instead of button for the suggested terms

Because the `href` is known so we can use links and not navigate programmatically with buttons.

* Fix search test

I had removed the `navigate` call because it wasn't needed for the
links, but it's also used programmatically to trigger the search in the
input.
 # Please enter the commit message for your changes. Lines
starting

* Fix SonarQube warning

* Create a context so the search dropdown can be closed

When terms, history, top search, products are clicked.

* Consider `closeSearchInputDropdown` to be null

Missed including this in the previous commit.

* Fix SonarQube warning

* Rename button from Clear to Clear History

* Adjust Loading text message spacing

* Fix link color for Top Search

* Fix grayish clock icon color in search history

* Limit search to 5 items per category

Noticed sometimes we were showing +5 and, apparently, the `first` pagination doesn't really filter the results.

* Also update the search input with the selected option

* Fix suggested terms link color

* Add the magnifying glass icon to suggested terms

* Run TS Hero extension to organize imports of touched files

* Better handle search input selections

Whenever a selection is made, be it from History, Top Search, Suggested Terms or Products, use the same logic:
- Add term to the history
- Send term to analytics
- Close dropdown
- Update the input with the term

* Do not show an empty gap when there are not results

* Move `SearchInputContext` and `useSearchInput` to `sdk/search/…`

* Fix warning about `SearchHistoryProps` type export

* Fix overflow-x when expanding the search on mobile

* Fix React key warning

* Debounce the search suggestions

* Add entries to the changelog

* Adjust `SuggestionsTopSearch` after the merged changes

* Fix Storybook search stories

* Simplify top search querying by using `useQuery`

Address [this suggestion](https://github.com/vtex-sites/gatsby.store/pull/67#discussion_r882857670).

* Remove debounce package in favor of using `useDeferredValue` from React

Address [this comment](https://github.com/vtex-sites/gatsby.store/pull/67#discussion_r882856990):

> can't we use React18 features like useDefferedValue to not need to implement a debounce function?

* Let TypeScript be aware of new types from React 18

So it doesn't complain when we use the new features (e.g., `useDeferredValue`).

* Let search history also record the path

Instead of using the generic search path for all the searched terms, along with the searched term, also store its path. This way, if a suggested product is selected, we store its name and its PDP path.

* Better organize by extracting `useTopSearch` to its own file

Address this comment:
https://github.com/vtex-sites/gatsby.store/pull/67#discussion_r887397506

* Better organize by extracting `useSuggestions` to its own file

Address this comment:
https://github.com/vtex-sites/gatsby.store/pull/67#discussion_r887397864

* Fix the missing search input (left) padding

Pointed out by this comment:
https://github.com/vtex-sites/gatsby.store/pull/67#pullrequestreview-992765808

> should we add padding-left inside the input?

What happened was that the padding was being overwritten by the styles when the search input is minimized.

---

Conflicts:
- CHANGELOG.md
- src/components/common/SearchInput/SearchInput.tsx
- src/components/search/History/SearchHistory.tsx
- src/components/search/SuggestionProductCard/SuggestionProductCard.tsx
- src/components/search/Suggestions/Suggestions.tsx
- src/components/search/Suggestions/SuggestionsTopSearch.tsx
- src/components/ui/Search/Suggestions/Suggestions.stories.tsx
- src/components/ui/Search/Suggestions/Suggestions.tsx