fix: Navigation links on mobile not being clickable
Noticed clicking on the links inside the navigation sidebar on mobile wasn't working. I figured it could be related to a recent z-index change from https://github.com/vtex-sites/nextjs.store/pull/212. The fix is to set the lower z-index only for desktop so the search dropdown is shown over the navigation bar. There's no need to set a z-index for mobile because the search dropdown and the navigation links are never displayed at the same time.